feat(Kan-202): input ok
