--- src/app/card-ready-to-use/travel-diary-card/travel-diary-card.component.css
@@ -1,24 +1,0 @@
-img {
-  width: 100%;
-  height: 100%;
-  -o-object-fit: cover;
-     object-fit: cover;
-}
-
-.informations {
-  flex: 1 1 auto;
-  display: flex;
-  flex-direction: column;
-  padding-left: 24px;
-  box-sizing: border-box;
-}
-.informations .title {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-.informations .description {
-  margin-top: 8px;
-  color: #666;
-  font-size: 0.875rem;
-}/*# sourceMappingURL=travel-diary-card.component.css.map */
