+++ src/app/components/card-ready-to-use/travel-diary-card/travel-diary-card.component.scss
@@ -3,6 +3,12 @@ img {
   height: 100%;
   object-fit: cover;
 }
+button {
+  all: unset;
+  &:hover {
+    opacity: 0.7;
+  }
+}
 
 .informations {
   flex: 1 1 auto;
