+++ src/app/components/card-ready-to-use/travel-diary-card/travel-diary-card.component.spec.ts
