+++ src/app/components/card-ready-to-use/visual-trip-card/visual-trip-card.component.spec.ts
